refactor signing functions
