+++ src/level_four/core/controller/controller_level_four.h
@@ -21,37 +21,12 @@
 #include "near.h"
 #include "solana.h"
 #include "tasks.h"
+#include "atecc_utils.h"
 
 #define DEFAULT_ATECC_RETRIES 5
 
-#define DEVICE_SERIAL_SIZE     32
 #define MAXIMUM_COIN_SUPPORTED (7 + 5)  // 5 for segwit support
 #define INDEX_SIZE             4
-
-typedef enum {
-  slot_0_unused       = 0U,
-  slot_1_unused       = 1U,
-  slot_2_auth_key     = 2U,
-  slot_3_nfc_pair_key = 3U,
-  slot_4_unused       = 4U,
-  slot_5_challenge    = 5U,
-  slot_6_io_key       = 6U,
-  slot_7_unused       = 7U,
-  slot_8_serial       = 8U,
-  slot_9_unused       = 9U,
-  slot_10_unused      = 10U,
-  slot_11_unused      = 11U,
-  slot_12_unused      = 12U,
-  slot_13_unused      = 13U,
-  slot_14_unused      = 14U,
-  slot_15_unused      = 15U
-} atecc_slot_define_t;
-
-typedef struct {
-  uint8_t device_serial[DEVICE_SERIAL_SIZE], retries;
-  ATCA_STATUS status;
-  ATCAIfaceCfg *cfg_atecc608a_iface;
-} atecc_data_t;
 
 extern atecc_data_t atecc_data;
